+++ src/http/pipeline.c
@@ -826,7 +826,7 @@ static void setPathInfo(MaConn *conn)
                 }
             }
             if (last) {
-                offset = alias->prefixLen + last - start;
+                offset = alias->prefixLen + (int) (last - start);
                 if (offset <= strlen(req->url)) {
                     pathInfo = &req->url[offset];
                     req->pathInfo = mprStrdup(req, pathInfo);
